Rename generate package to golang
