+++ internal/generate/golang/code.go
@@ -1,4 +1,4 @@
-package generate
+package golang
 
 import (
 	"regexp"
