+++ internal/generate/golang/golang.go
@@ -1,4 +1,4 @@
-package generate
+package golang
 
 import (
 	"bytes"
