+++ internal/generate/golang/golang_test.go
@@ -1,4 +1,4 @@
-package generate
+package golang
 
 import (
 	"os"
