+++ internal/generate/golang/templates/errors.go.tmpl
