+++ internal/generate/golang/templates/input.go.tmpl
