+++ internal/generate/golang/templates/lexer.go.tmpl
