+++ internal/generate/golang/templates/parser.go.tmpl
