+++ internal/generate/golang/templates/stack.go.tmpl
