+++ internal/generate/golang/templates/types.go.tmpl
